Fix searching
Fix searching of database.  It's not done in a "proper" way, but it uses string formatting for the search field but as a caution it checks possible fields against a list of valid ones.

I feel this is a reasonable approach because it simply will not work any other way.  Due to the design of the pysqlite3 module, placeholders cannot be used for the column names.  Because I was able to mitigate the chances of a SQL injection attack by checking any input against a list of valid ones it should remain secure.

Signed-off-by: Braxton Schafer <braxton.schafer@gmail.com>