Improve Online Search Mechanism
Improve the online search mechanism so the results returned are more correct and usable.  Although it is not "less hardcoded", it accomplishes what I was hoping to accomplish.

Remove my branding from the top of the file that was auto-added by Eclipse.

Add ASCII art of Pyberry logo at the start of the program.