Backport to Python 2.7
This is a major release, as I've decided to move to Python 2.7 (from Python 3) due to module reasons.  This commit sees all of the code transformed into Python 2.x syntax.

Additionally, due to the change, I could fix several outstanding issues.  This fixes #2 because the whole lookup class is rewritten to use the Google API library module.  Although I only added search by title, it should be very straightforward to be add any other attribute, so this closes #3.

It also fixes #10, although that doesn't mean that multiple authors are completely supported, only that there's the option in the future.

Signed-off-by: Braxton Schafer <braxton.schafer@gmail.com>