Rewrite Lookup to use OpenLibrary and its API
After a lot of struggles with poor documentation on the Google Books API and its client, I decided to move to OpenLibrary for now, in the hopes it will work better.  This is the rewrite of the Lookup class to use OpenLibrary.  I have not yet modified the main method or done any testing.