Change mid-process exit to return
If the user wanted to exit in the middle of, say, adding a book or searching for existing books, it was done by throwing and catching an exception.  I realized that’s a pretty weird way of doing it, so I switched to returning values instead.  Makes more sense.

Also it’s not stated in the UX, but if either the search field or search terms are blank (in searching for local books), it will gently return you to the main screen.